+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [#11](https://github.com/dblock/obcd/pull/11): Fix: incorrect violation count with multiple violations from a single check - [@interstateone](https://github.com/interstateone).
 * [#11](https://github.com/dblock/obcd/pull/11): Fix: HeaderStyle incorrectly detects blank line after copyright - [@interstateone](https://github.com/interstateone).
 * [#16](https://github.com/dblock/obcd/pull/16): Fix: broken links in CONTRIBUTING.md - [@w5mith](https://github.com/w5mith).
+* [#15](https://github.com/dblock/obcd/pull/15): Added [SpectaFocus](docs/checks/SpectaFocus.md) check - [@w5mith](https://github.com/w5mith).
 
 ### 0.1.0 (3/28/2014)